refactor: organize sub rows into own dir
