+++ ui/src/components/timeline-tree/QueryResourceTree.test.tsx
@@ -20,7 +20,7 @@ import {
   OPERATOR_TIMELINE_ROW_TYPE,
   type TreeTableItem,
 } from '@quent/components';
-import type { ResourceTimelineSubRow } from './subRow';
+import type { ResourceTimelineSubRow } from './sub-rows';
 
 // ---------------------------------------------------------------------------
 // Mock heavy/visual dependencies so tests run without a real browser/canvas
