+++ ui/src/components/timeline-tree/sub-rows/index.ts
@@ -1,0 +1,12 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export {
+  createSyntheticSubRow,
+  createTimelineSubRow,
+  mapTreeItems,
+  TimelineSubRowLabel,
+  type ResourceTimelineSubRow,
+} from './subRow';
+export { createLongEntitiesTimelineSubRow } from './LongEntitiesTimelineSubRow';
+export { createOperatorGanttTimelineSubRow } from './OperatorGanttTimelineSubRow';
